+++ cmd/substreams/gui.go
@@ -44,7 +44,7 @@ Stream module output from a given package on a remote endpoint. The manifest is 
 'substreams.yaml' in current working directory if nothing entered. You may enter a directory that contains a 'substreams.yaml'
 file in place of '<manifest_file>, or a link to a remote .spkg file, using urls gs://, http(s)://, ipfs://, etc.'.
 
-You can also use substreams gui my-package@v0.1.0 to specify a specific version of the package. This will fetch it from 
+You can also use substreams gui my-package@v0.1.0 to specify a specific version of the package. This will fetch it from
 the Substreams registry at https://substreams.dev
 `)
 
@@ -163,6 +163,7 @@ func runGui(cmd *cobra.Command, args []string) (err error) {
 		authToken, authType,
 		sflags.MustGetBool(cmd, "insecure"),
 		sflags.MustGetBool(cmd, "plaintext"),
+		"substreams_gui",
 	)
 
 	homeDir, err := os.UserHomeDir()